fix filter on home page by adding a separated component for it
